+++ src/tests/mod.rs
@@ -698,3 +698,6 @@ where
 
     addr
 }
+
+pub(crate) fn assert_send<T: Send>() {}
+pub(crate) fn assert_sync<T: Sync>() {}